requirements.txt
pyodbc==5.1.0
tabulate==0.9.0
openai==1.12.0
langchain==0.1.6
langchain-community==0.0.20
langchain-core==0.1.23
langchain-experimental==0.0.49
langchain-openai==0.0.5
pandas==2.2.2
SQLAlchemy==2.0.30
pandas==2.2.2
python-dotenv==1.0.1
streamlit
langdetect